pyproject.toml
[tool.poetry]
name = "colors"
version = "0.1.0"
description = ""
authors = ["Alan Reyes <alan.reyes@mongodb.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
pymongo = {extras = ["dns"], version = "^4.3.2"}
python-dotenv = "^0.21.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
